[IMP] {website_}event: add answers in exemple reports and demo data
As foldable badges and full page tickets may contain answers
to simple choice questions as tags, if any, add them in the
'example' badge and full page ticket to indicate they will
appear there.

Show the first answer of any such question set on the event
as a tag. This way, the answers will not be hard coded but
will match the event data.

Also give some simple_choice answers on the demo registrations
of Openwood Collection Reveal, extending the events of which
attendees have answer tag(s) displayed on their ticket or foldable
badge.

Code is simplified to have single blocks.

Task-4048204

closes odoo/odoo#173310

Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>